+++ pkg/kudoctl/util/repo/index_test.go
@@ -177,7 +177,7 @@ func TestMapPackageFileToPackageVersion(t *testing.T) {
 		Maintainers:       []*v1beta1.Maintainer{{Name: "Ken Sipe"}},
 		URL:               "http://kudo.dev/kafka",
 	}
-	pf := packages.PackageFiles{
+	pf := packages.Files{
 		Operator: &o,
 	}
